+++ tests/Unit/CustomValueTest.php
@@ -11,7 +11,8 @@ class CustomValueTest extends TableListTestCase
     public function testSetIsCustomValueAttribute()
     {
         $table = app(TableList::class)->setModel(User::class);
-        $closure = function($entity, $column) { };
+        $closure = function ($entity, $column) {
+        };
         $table->addColumn('name')->isCustomValue($closure);
         $this->assertEquals($closure, $table->columns->first()->customValueClosure);
     }